Make schema Ruby-friendly
PyKwalify is intended as a port of
http://www.kuwata-lab.com/kwalify/
so it seems we should keep out input schema interoperable
when possible.

Before this commit:

```
$ pykwalify -s opencontrol-component-kwalify-schema.yaml -d component-example.yaml
 INFO - validation.valid
$ kwalify -f opencontrol-component-kwalify-schema.yaml component-example.yaml
ERROR: [/mapping/name/required] 'True': not a boolean.
```

After

```
$ pykwalify -s opencontrol-component-kwalify-schema.yaml -d component-example.yaml
 INFO - validation.valid
$ kwalify -f opencontrol-component-kwalify-schema.yaml component-example.yaml
component-example.yaml#0: valid.
```

As there doesn't seem to be any downside from changing 'True' to 'true' how about we go for it?